+++ src/config/config.ts
@@ -32,11 +32,12 @@ export namespace customConfig {
 		limit: string   // limit unit is Wei
 	}
 
-	class Watchdog {
+	export class WatchdogConfig {
 		enable: boolean
 		mailer: Mailer
 		callback: string
 		period: number
+		blockTime: number
 		addressList: WatchedAddress[]
 	}
 
@@ -95,7 +96,7 @@ export namespace customConfig {
 	export class CustomConfig {
 		apiResponseCode: any
 		tx: TxConfig
-		watchdog?: Watchdog
+		watchdog?: WatchdogConfig
 		events?: Events
 		database?: Database
 	}
@@ -116,7 +117,7 @@ export namespace customConfig {
 	}
 
 	// GetWatchdog returns watchdog config
-	export function GetWatchdog(): Watchdog {
+	export function GetWatchdog(): WatchdogConfig {
 		if (customConfig.watchdog) {
 			return customConfig.watchdog;
 		}