+++ src/mailer/index.ts
@@ -1,2 +1,3 @@
 export * as mailer from './mailer';
-export {Template} from './template';
+export {Mailer} from './mailer';
+export {MailContent, MailQueue} from './types';