+++ src/watchdog/params.ts
@@ -1,5 +1,29 @@
+// Default check address list job concurrency
+export const DefaultCheckAddressListJobConcurrency = 1;
+
+// Default check address balance job concurrency
+export const DefaultCheckAddressBalanceJobConcurrency = 1;
+
 // Default loop interval (unit: ms)
 export const DefaultLoopInterval = 3000;
+
+// Default balance cache max limit
+export const DefaultBalanceCacheMaxLimit = 1;
+
+// Retry times
+export const DefaultRetryTimes = 5;
+
+// Retry min interval (unit: second)
+export const DefaultRetryMinInterval = 3;
+
+// Retry max interval (unit: second)
+export const DefaultRetryMaxInterval = DefaultRetryMinInterval * 3;
+
+// Default interval for checking mail queue
+export const DefaultCheckMailQueueInterval = 3000;
+
+// Default send alert mail job concurrency
+export const DefaultSendAlertMailJobConcurrency = 1;
 
 // Default alert type
 export const enum DefaultAlertType {