+++ src/watchdog/watchdog.ts
@@ -1,62 +1,111 @@
-import {auditor, log} from '@jovijovi/pedrojs-common';
-import {BigNumber, utils} from 'ethers';
 import * as util from 'util';
-import fastq, {queueAsPromised} from 'fastq';
 import got from 'got';
-import {Queue, retry} from '@jovijovi/pedrojs-common/util';
-import fs from 'fs';
-import {MailContent} from '../mailer/mailer';
-import {IsAlert} from './rules';
-import {customConfig} from '../config';
-import {mailer, Template} from '../mailer';
+import {BigNumber, utils} from 'ethers';
+import fastq, {queueAsPromised} from 'fastq';
+import {auditor, log} from '@jovijovi/pedrojs-common';
+import {Queue} from '@jovijovi/pedrojs-common/util';
 import {network} from '@jovijovi/ether-network';
-import {DefaultAlertMailTemplate, DefaultAlertType, DefaultLoopInterval} from './params';
-import {BalanceCache} from './types';
+import {customConfig} from '../config';
+import {
+	DefaultAlertType,
+	DefaultCheckAddressBalanceJobConcurrency,
+	DefaultCheckAddressListJobConcurrency,
+	DefaultLoopInterval
+} from './params';
+import {AlertGeneratorParams, Balance, CheckAddressBalanceJobParams, FuncAlertGenerator} from './types';
+import {IsAlert} from './rules';
+import {BuildBalanceAlertMailContent} from './template';
+import {GetPastBalance, SetCache} from './cache';
+import {AlertMailer, GenHtmlMail, InitAlertMailer} from './mail';
+import {RetryGetBalance} from './utils';
 
-const blockQueue = new Queue<number>();     // Block queue (ASC, FIFO)
-const checkBalanceJob: queueAsPromised<number> = fastq.promise(checkAddressList, 1);    // Job: Check balance
-const balanceCacheMaxLimit = 1;             // Balance cache max limit
-const balanceCache: BalanceCache<any> = {}; // Balance cache
+// Block queue (ASC, FIFO)
+const blockQueue = new Queue<number>();
 
+// Job: Check address list
+const checkAddressListJob: queueAsPromised<number> = fastq.promise(checkAddressList, DefaultCheckAddressListJobConcurrency);
+
+// Job scheduler: Check address balance
+const checkAddressBalanceJobScheduler: Map<string, queueAsPromised<CheckAddressBalanceJobParams>> = new Map();
+
+// Run watchdog
 export function Run() {
+	const [conf, ok] = init();
+	if (!ok) {
+		return;
+	}
+
+	// Listen block event
+	listenBlockEvent(conf.period);
+
+	// Schedule processing job
+	setInterval(() => {
+		const length = blockQueue.Length();
+		if (length === 0) {
+			return;
+		}
+
+		// Try to catch up the latest block
+		for (let i = 0; i < length; i++) {
+			const blockNumber = blockQueue.Shift();
+			if (!blockNumber) {
+				return;
+			}
+
+			checkAddressListJob.push(blockNumber).catch((err) => log.RequestId().error(err));
+		}
+	}, conf.blockTime ? conf.blockTime : DefaultLoopInterval);
+
+	log.RequestId().info("Watchdog is running...");
+}
+
+// Init watchdog
+function init(): [customConfig.WatchdogConfig, boolean] {
 	// Check config
-	const watchdogConf = customConfig.GetWatchdog();
-	if (!watchdogConf) {
+	const conf = customConfig.GetWatchdog();
+	if (!conf) {
 		log.RequestId().info('No watchdog configuration, skipped.');
 		return;
-	} else if (!watchdogConf.enable) {
+	} else if (!conf.enable) {
 		log.RequestId().info('Watchdog disabled.');
 		return;
 	}
 
-	log.RequestId().info("Watchdog is running...");
+	log.RequestId().info("Watchdog config=\n%s", JSON.stringify(conf, undefined, 4));
 
+	// Check params
+	auditor.Check(conf.period > 0, 'invalid period');
+	auditor.Check(conf.blockTime > 0, 'invalid blockTime');
+
+	// Check address list
+	if (!conf.addressList) {
+		throw new Error(`invalid addressList`);
+	}
+	for (const watchedAddress of conf.addressList) {
+		auditor.Check(utils.isAddress(watchedAddress.address), 'invalid address');
+	}
+
+	// Init alert mailer
+	InitAlertMailer();
+
+	return [conf, true];
+}
+
+// Listen block event
+function listenBlockEvent(period: number) {
 	const provider = network.MyProvider.Get();
 	let nextBlock = 0;
 	provider.on('block', (blockNumber) => {
 		if (blockNumber >= nextBlock) {
 			blockQueue.Push(blockNumber);
-			nextBlock = blockNumber + watchdogConf.period;
+			nextBlock = blockNumber + period;
 		}
 	})
-
-	setInterval(() => {
-		if (blockQueue.Length() == 0) {
-			return;
-		}
-
-		const blockNumber = blockQueue.Shift();
-		if (!blockNumber) {
-			return;
-		}
-
-		checkBalanceJob.push(blockNumber).catch((err) => log.RequestId().error(err));
-	}, DefaultLoopInterval);
 }
 
+// Check address list
 async function checkAddressList(blockNumber: number): Promise<void> {
 	auditor.Check(blockNumber >= 0, 'invalid blockNumber');
-	// log.RequestId().debug("# Check # BlockNumber: ", blockNumber);
 
 	// Get address list
 	const addressList = customConfig.GetWatchdog().addressList;
@@ -67,174 +116,60 @@ async function checkAddressList(blockNumber: number): Promise<void> {
 	// Check address list
 	for (let i = 0; i < addressList.length; i++) {
 		try {
-			await checkAddressBalance(addressList[i], blockNumber);
+			const address = addressList[i].address;
+			auditor.Check(utils.isAddress(address), 'invalid address');
+
+			if (!checkAddressBalanceJobScheduler.has(address)) {
+				checkAddressBalanceJobScheduler.set(address, fastq.promise(checkAddressBalance, DefaultCheckAddressBalanceJobConcurrency));
+			}
+
+			checkAddressBalanceJobScheduler.get(address).push({
+				watchedAddress: addressList[i],
+				blockNumber: blockNumber,
+			}).catch((err) => log.RequestId().error(err));
 		} catch (e) {
-			log.RequestId().error("Check address balance(%s) failed, blockNumber=%d, error=%o", addressList[i], blockNumber, e);
+			log.RequestId().error("Check balance of address(%s)@block(%d) failed, error=%o",
+				addressList[i], blockNumber, e);
 		}
 	}
 }
 
-// Combination cache key
-function combinationCacheKey(keys: string[]): string {
-	return keys.toString();
-}
+// Check address balance
+async function checkAddressBalance(opts: CheckAddressBalanceJobParams) {
+	// Check params
+	auditor.Check(utils.isAddress(opts.watchedAddress.address), 'invalid address');
+	auditor.Check(opts.blockNumber >= 0, 'invalid blockNumber');
 
-// Set cache
-function setCache(address: string, rule: string, balance: BigNumber) {
-	const key = combinationCacheKey([address, rule]);
-	if (!balanceCache[key]) {
-		balanceCache[key] = new Queue(balanceCacheMaxLimit);
+	// Get previous balance & current balance
+	const balance = await getBalance(opts);
+
+	// Check if is alert
+	const [isAlert, alertType] = IsAlert(opts.watchedAddress, balance);
+	if (!isAlert) {
+		return;
 	}
 
-	balanceCache[key].Push(balance);
-}
+	// Generate alert
+	const [msg, subject] = alertType === DefaultAlertType.BalanceReachLimit ?
+		balanceReachLimitAlert({
+			balance: balance,
+			blockNumber: opts.blockNumber,
+			watchedAddress: opts.watchedAddress
+		})()
+		: balanceChangesAlert({
+			balance: balance,
+			blockNumber: opts.blockNumber,
+			watchedAddress: opts.watchedAddress,
+		})();
 
-// Get past balance from cache
-function getPastBalance(address: string, rule: string): BigNumber {
-	const key = combinationCacheKey([address, rule]);
-	if (!balanceCache[key]) {
-		balanceCache[key] = new Queue(balanceCacheMaxLimit);
-	}
+	// Add alert mail to queue
+	AlertMailer().Queue().Push(BuildBalanceAlertMailContent({
+		subject: subject,
+		html: GenHtmlMail(alertType, msg),
+	}));
 
-	return balanceCache[key].First();
-}
-
-async function checkAddressBalance(watchedAddress: customConfig.WatchedAddress, blockNumber: number) {
-	auditor.Check(utils.isAddress(watchedAddress.address), 'invalid address');
-	auditor.Check(blockNumber >= 0, 'invalid blockNumber');
-
-	const provider = network.MyProvider.Get();
-
-	let balancePast = getPastBalance(watchedAddress.address, watchedAddress.rule);
-	if (!balancePast) {
-		balancePast = await retry.Run(async (): Promise<BigNumber> => {
-			return await provider.getBalance(watchedAddress.address, blockNumber < 1 ? 0 : blockNumber - 1);
-		});
-		setCache(watchedAddress.address, watchedAddress.rule, balancePast);
-	}
-
-	const balanceNow = await retry.Run(async (): Promise<BigNumber> => {
-		return await provider.getBalance(watchedAddress.address, blockNumber);
-	});
-	if (!balanceNow.eq(balancePast)) {
-		setCache(watchedAddress.address, watchedAddress.rule, balanceNow);
-	}
-
-	const [isAlert, alertType] = IsAlert(watchedAddress, {
-		Now: balanceNow,
-		Past: balancePast,
-	});
-	if (isAlert) {
-		// Function to generate BalanceReachLimit alert
-		const genBalanceReachLimitAlert = (): [any, string] => {
-			const alertMsg = util.format("Address(%s) balance (%s) reaches limit (%s) at blockNumber(%d) [Chain:%s Network:%s ChainId:%s]",
-				watchedAddress.address, utils.formatEther(balanceNow), utils.formatEther(watchedAddress.limit), blockNumber,
-				network.GetDefaultNetwork().chain, network.GetDefaultNetwork().network, network.GetChainId());
-
-			log.RequestId().info("***** ALERT ***** %s", alertMsg);
-
-			const msg = {
-				address: watchedAddress.address,
-				rule: watchedAddress.rule,
-				balanceNow: utils.formatEther(balanceNow),
-				limit: utils.formatEther(watchedAddress.limit),
-				addressUrl: util.format("%s/address/%s", network.GetBrowser(), watchedAddress.address),
-				blockNumber: blockNumber.toString(),
-				chain: network.GetDefaultNetwork().chain,
-				network: network.GetDefaultNetwork().network,
-				chainId: network.GetChainId(),
-			}
-			const subject = util.format("%s reaches limit (%s) @%d", watchedAddress.address, utils.formatEther(watchedAddress.limit), blockNumber);
-			return [msg, subject];
-		}
-
-		// Function to generate BalanceChanges alert
-		const genBalanceChangesAlert = (): [any, string] => {
-			const alertMsg = util.format("Address(%s) balance changed from (%s) to (%s) at blockNumber(%d) [Chain:%s Network:%s ChainId:%s]",
-				watchedAddress.address, utils.formatEther(balancePast), utils.formatEther(balanceNow), blockNumber,
-				network.GetDefaultNetwork().chain, network.GetDefaultNetwork().network, network.GetChainId());
-
-			log.RequestId().info("***** ALERT ***** %s", alertMsg);
-
-			const changed = balanceNow.sub(balancePast);
-			const plusSign = changed.gt(BigNumber.from(0)) ? '+' : '';
-			const balanceChanged = plusSign + utils.formatEther(balanceNow.sub(balancePast));
-			const msg = {
-				address: watchedAddress.address,
-				rule: watchedAddress.rule,
-				balanceChanged: balanceChanged,
-				balanceNow: utils.formatEther(balanceNow),
-				balancePast: utils.formatEther(balancePast),
-				addressUrl: util.format("%s/address/%s", network.GetBrowser(), watchedAddress.address),
-				blockNumber: blockNumber.toString(),
-				chain: network.GetDefaultNetwork().chain,
-				network: network.GetDefaultNetwork().network,
-				chainId: network.GetChainId(),
-			}
-			const subject = util.format("%s (%s @%d)", watchedAddress.address, balanceChanged, blockNumber);
-			return [msg, subject];
-		}
-
-		// Generate alert
-		const [msg, subject] = alertType === DefaultAlertType.BalanceReachLimit ? genBalanceReachLimitAlert() : genBalanceChangesAlert();
-
-		// Send alert
-		await sendAlert(Template.BalanceAlertMailContent({
-			subject: subject,
-			html: genHtmlMail(alertType, msg),
-		}));
-
-		// Callback (optional)
-		await callback(msg);
-	}
-}
-
-function genHtmlMail(alertType: DefaultAlertType, arg: any): string {
-	let template = '';
-
-	switch (alertType) {
-		case DefaultAlertType.BalanceReachLimit:
-			template = fs.readFileSync(DefaultAlertMailTemplate.BalanceReachLimit.Path, 'utf8');
-			template = template.replace(/\${address}/gi, arg.address);
-			template = template.replace(/\${rule}/gi, arg.rule);
-			template = template.replace(/\${balanceNow}/gi, arg.balanceNow);
-			template = template.replace(/\${limit}/gi, arg.limit);
-			template = template.replace(/\${addressUrl}/gi, arg.addressUrl);
-			template = template.replace(/\${blockNumber}/gi, arg.blockNumber);
-			template = template.replace(/\${chain}/gi, arg.chain);
-			template = template.replace(/\${network}/gi, arg.network);
-			template = template.replace(/\${chainId}/gi, arg.chainId);
-			return template;
-
-		case DefaultAlertType.BalanceChanges:
-			template = fs.readFileSync(DefaultAlertMailTemplate.BalanceChanges.Path, 'utf8');
-			template = template.replace(/\${address}/gi, arg.address);
-			template = template.replace(/\${rule}/gi, arg.rule);
-			template = template.replace(/\${balanceNow}/gi, arg.balanceNow);
-			template = template.replace(/\${balancePast}/gi, arg.balancePast);
-			template = template.replace(/\${balanceChanged}/gi, arg.balanceChanged);
-			template = template.replace(/\${addressUrl}/gi, arg.addressUrl);
-			template = template.replace(/\${blockNumber}/gi, arg.blockNumber);
-			template = template.replace(/\${chain}/gi, arg.chain);
-			template = template.replace(/\${network}/gi, arg.network);
-			template = template.replace(/\${chainId}/gi, arg.chainId);
-			return template;
-	}
-
-	return undefined;
-}
-
-// sendAlert send alert by mail
-async function sendAlert(mailContent: MailContent) {
-	try {
-		if (!customConfig.GetWatchdog().mailer) {
-			return;
-		}
-
-		await mailer.Send(customConfig.GetWatchdog().mailer, mailContent);
-	} catch (e) {
-		log.RequestId().error("SendAlert failed, error=", e);
-	}
+	// Callback (optional)
+	await callback(msg);
 }
 
 // callback
@@ -253,4 +188,84 @@ async function callback(msg: any) {
 	} catch (e) {
 		log.RequestId().error("Callback failed, error=", e);
 	}
+}
+
+// Get previous balance & current balance
+async function getBalance(opts: CheckAddressBalanceJobParams): Promise<Balance> {
+	// Get previous balance
+	let balancePrevious = GetPastBalance(opts.watchedAddress.address, opts.watchedAddress.rule);
+	if (!balancePrevious) {
+		log.RequestId().trace("Getting previous balance of address(%s)@block(%d)...",
+			opts.watchedAddress.address, opts.blockNumber - 1);
+		balancePrevious = await RetryGetBalance(opts.watchedAddress.address, opts.blockNumber < 1 ? 0 : opts.blockNumber - 1)
+		SetCache(opts.watchedAddress.address, opts.watchedAddress.rule, balancePrevious);
+	}
+
+	// Get current balance
+	log.RequestId().trace("Getting current balance of address(%s)@block(%d)...",
+		opts.watchedAddress.address, opts.blockNumber);
+	const balanceCurrent = await RetryGetBalance(opts.watchedAddress.address, opts.blockNumber);
+	if (!balanceCurrent.eq(balancePrevious)) {
+		SetCache(opts.watchedAddress.address, opts.watchedAddress.rule, balanceCurrent);
+	}
+
+	return {
+		Previous: balancePrevious,
+		Current: balanceCurrent,
+	};
+}
+
+// Function to generate BalanceReachLimit alert
+function balanceReachLimitAlert(opts: AlertGeneratorParams): FuncAlertGenerator {
+	return (): [any, string] => {
+		const alertMsg = util.format("Address(%s) balance (%s) reaches limit (%s) at blockNumber(%d) [Chain:%s Network:%s ChainId:%s]",
+			opts.watchedAddress.address, utils.formatEther(opts.balance.Current), utils.formatEther(opts.watchedAddress.limit), opts.blockNumber,
+			network.GetDefaultNetwork().chain, network.GetDefaultNetwork().network, network.GetChainId());
+
+		log.RequestId().info("***** ALERT ***** %s", alertMsg);
+
+		const msg = {
+			address: opts.watchedAddress.address,
+			rule: opts.watchedAddress.rule,
+			balanceCurrent: utils.formatEther(opts.balance.Current),
+			limit: utils.formatEther(opts.watchedAddress.limit),
+			addressUrl: util.format("%s/address/%s", network.GetBrowser(), opts.watchedAddress.address),
+			blockNumber: opts.blockNumber.toString(),
+			chain: network.GetDefaultNetwork().chain,
+			network: network.GetDefaultNetwork().network,
+			chainId: network.GetChainId(),
+		}
+		const subject = util.format("%s reaches limit (%s) @%d",
+			opts.watchedAddress.address, utils.formatEther(opts.watchedAddress.limit), opts.blockNumber);
+		return [msg, subject];
+	};
+}
+
+// Function to generate BalanceChanges alert
+function balanceChangesAlert(opts: AlertGeneratorParams): FuncAlertGenerator {
+	return (): [any, string] => {
+		const alertMsg = util.format("Address(%s) balance changed from (%s) to (%s) at blockNumber(%d) [Chain:%s Network:%s ChainId:%s]",
+			opts.watchedAddress.address, utils.formatEther(opts.balance.Previous), utils.formatEther(opts.balance.Current), opts.blockNumber,
+			network.GetDefaultNetwork().chain, network.GetDefaultNetwork().network, network.GetChainId());
+
+		log.RequestId().info("***** ALERT ***** %s", alertMsg);
+
+		const changed = opts.balance.Current.sub(opts.balance.Previous);
+		const plusSign = changed.gt(BigNumber.from(0)) ? '+' : '';
+		const balanceChanged = plusSign + utils.formatEther(opts.balance.Current.sub(opts.balance.Previous));
+		const msg = {
+			address: opts.watchedAddress.address,
+			rule: opts.watchedAddress.rule,
+			balanceChanged: balanceChanged,
+			balanceCurrent: utils.formatEther(opts.balance.Current),
+			balancePrevious: utils.formatEther(opts.balance.Previous),
+			addressUrl: util.format("%s/address/%s", network.GetBrowser(), opts.watchedAddress.address),
+			blockNumber: opts.blockNumber.toString(),
+			chain: network.GetDefaultNetwork().chain,
+			network: network.GetDefaultNetwork().network,
+			chainId: network.GetChainId(),
+		}
+		const subject = util.format("%s (%s @%d)", opts.watchedAddress.address, balanceChanged, opts.blockNumber);
+		return [msg, subject];
+	};
 }